Initial check-in
This code implements a Grace hash join to simulate the following query :
 select count(*), sum(L_EXTENDEDPRICE) from ORDERS, LINEITEM  where
L_ORDERKEY = O_ORDERKEY
 and O_ORDERDATE between 1995-01-02 and 1996-01-02;

The code uses flat files from TPCH.

+++ src/spillingHashJoin/TempFileWriter.java
@@ -1,13 +1,12 @@
-package HashToDisk;
+package spillingHashJoin;
 import java.io.BufferedWriter;
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
-import java.io.UnsupportedEncodingException;
 import java.nio.charset.Charset;
-import java.nio.file.*;
+import java.nio.file.Path;
 
-public class SpillFileHandler {
+public class TempFileWriter {
 
 	public String getFileName() {
 		return fileName;
@@ -17,13 +16,13 @@ public class SpillFileHandler {
 		return filePath;
 	}
 
-	public SpillFileHandler(String fileName) throws IOException
+	public TempFileWriter(String fileName, boolean deleteIfExists) throws IOException
 	{
 		this.fileName = fileName;
 		fileToWriteTo = new File(fileName);
 		filePath = fileToWriteTo.toPath();
 		
-		if (fileToWriteTo.exists())
+		if (fileToWriteTo.exists() && deleteIfExists)
 		{
 			fileToWriteTo.delete();
 		}
